refactor(mempool): extract txTTLExpired helper; use uint for cache params
Extract TTL-block eviction logic in admitter.go into txTTLExpired,
mirroring the txTimedout pattern. Mirrors reviewer suggestion to
compress the inline block into a named helper.

Change NewEncoderCache/NewDecodeCache size+maxTxBytes from int to uint:
negative values are now impossible at the type level. 0 still falls back
to cmdcfg defaults. Update callers in app.go (uint casts) and remove
the now-unreachable negative test case in decoder_cache_test.go.

+++ app/app.go
@@ -388,7 +388,7 @@ func New(
 			panic(fmt.Errorf("%s (%d) must not exceed %s (%d)", FlagTxCacheMaxTxBytes, maxTxBytes, FlagMempoolMaxTxBytes, mempoolMaxTxBytes))
 		}
 		logger.Info("tx encode/decode cache enabled", "size", txCacheSize, "max-tx-bytes", maxTxBytes)
-		activeDecoder = cronosmempool.NewCachingDecoder(txDecoder, cronosmempool.NewDecodeCache(txCacheSize, maxTxBytes))
+		activeDecoder = cronosmempool.NewCachingDecoder(txDecoder, cronosmempool.NewDecodeCache(uint(txCacheSize), uint(maxTxBytes)))
 	}
 	eip712.SetEncodingConfig(encodingConfig)
 
@@ -493,7 +493,7 @@ func New(
 
 		var encCache *cronosmempool.EncoderCache
 		if mempoolType == cronosmempool.TypeApp && txCacheSize >= 0 {
-			encCache = cronosmempool.NewEncoderCache(txCacheSize, maxTxBytes)
+			encCache = cronosmempool.NewEncoderCache(uint(txCacheSize), uint(maxTxBytes))
 		}
 
 		// baseFee gate source for the app-path selector; proposalFee is assigned
